:zap: split timetables
