+++ src/components/schedule/Zenyasai.astro
@@ -1,0 +1,331 @@
+---
+import Talk from "./Talk.astro";
+import MiniPanel from "../Layouts/MiniPanel.astro";
+import ArrowUpRightFromSquare from '../icons/arrow-up-right-from-square.svg';
+---
+<style>
+  .schedule {
+    padding: 24px 0;
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+
+    .ann {
+      .ann-title {
+        h3{
+          font-size: 20px;
+          font-weight: 600;
+          line-height: 30px;
+          text-align: left;
+        }
+
+        padding-bottom: 16px;
+        border-bottom: 1px solid #aeaeb2;
+
+        margin-bottom: 16px;
+      }
+
+      .ann-body {
+        p {
+          font-size: 14px;
+          font-weight: 300;
+          line-height: 32px;
+          text-align: left;
+          text-decoration: none;
+        }
+
+        .ann-detail{
+          display: flex;
+          justify-content: center;
+          margin-top: 16px;
+
+          .ann-button {
+            a {
+              font-size: 16px;
+              font-weight: 600;
+              line-height: 24px;
+              text-align: left;
+              text-decoration: none;
+            }
+
+            border: 2px solid #AEAEB2;
+            padding: 16px 24px;
+            border-radius: 4px;
+            background: #FFFFFF;
+
+            display: flex;
+            gap: 8px;
+          }
+        }
+      }
+    }
+
+    .timetable{
+      overflow-x: auto;
+      -webkit-overflow-scrolling: touch;
+
+      table thead:first-child tr:first-child th:first-child {
+        border-top-left-radius: 24px;
+        border-right: 1px solid #AEAEB2;
+      }
+      table thead:first-child tr:first-child th:last-child {
+        border-top-right-radius: 24px;
+        border-left: 1px solid #AEAEB2;
+      }
+
+      table tbody:first-child tr:first-child td:first-child {
+        border-top-left-radius: 24px;
+        border-top: none !important;
+        border-right: 1px solid #AEAEB2;
+      }
+      table tbody:first-child tr:first-child td:last-child {
+        border-top-right-radius: 24px;
+        border-top: none !important;
+      }
+
+      table {
+        width: 100%;
+        min-width: 1028px;
+
+        border-spacing: 0;
+        border-radius: 24px;
+        border: 1px solid #AEAEB2;
+
+        .tt-no-left-line {
+          border-left: none !important;
+        }
+
+        tbody tr:last-child td:first-child {
+          border-bottom-left-radius: 24px;
+        }
+        tbody tr:last-child td:last-child {
+          border-bottom-right-radius: 24px;
+        }
+
+        tbody tr td:first-child {
+          border-right: 1px solid #AEAEB2;
+          background-color: #F5F5F5;
+        }
+
+        tbody tr td:nth-child(3) {
+          border-left: 1px solid #AEAEB2;
+        }
+
+        td {
+          border-top: 1px solid #AEAEB2;
+          padding: 24px;
+        }
+
+        th {
+          font-family: "Futura", "Jost", sans-serif;
+          font-size: 20px;
+          font-weight: 500;
+          line-height: 27px;
+          text-align: center;
+
+          padding: 24px 16px;
+        }
+      }
+      @media screen and (width <= 720px) {
+        table {
+          min-width: 688px;
+        }
+      }
+      @media screen and (width <= 520px) {
+        table {
+          min-width: 488px;
+        }
+      }
+      @media screen and (width <= 320px) {
+        table {
+          min-width: 320px;
+        }
+      }
+
+      .tt-4 {
+        color: #fff;
+        background: #DD451D;
+      }
+
+      .tt-3 {
+        color: #fff;
+        background: #000;
+      }
+
+      .tt-time, .tt-timerange {
+        font-family: "Futura", "Jost", sans-serif;
+        font-size: 16px;
+        font-weight: 500;
+        line-height: 21.25px;
+        text-align: center;
+
+        white-space: nowrap;
+        width: 1%;
+      }
+
+      .tt-event {
+        font-size: 20px;
+        font-weight: 600;
+        line-height: 32px;
+        text-align: left;
+      }
+
+      .tt-4, .tt-3, .tt-event, .tt-talk {
+        width: 50%;
+      }
+    }
+  }
+</style>
+
+<div class="schedule">
+  <div class="ann">
+    <MiniPanel>
+      <div>
+        <div class="ann-title">
+          <h3>東京Ruby会議12 前夜祭</h3>
+        </div>
+        <div class="ann-body">
+          <p>
+            前夜祭には当日チケットを購入した方のみご参加いただけます。詳細は connpass をご確認ください。
+          </p>
+          <div class="ann-detail">
+            <div class="ann-button">
+              <a
+                href="https://connpass.com/event/339170/"
+                target="_blank"
+                rel="noreferrer"
+              >詳細を確認する</a>
+              <img src={ArrowUpRightFromSquare.src} alt="link icon"/>
+            </div>
+          </div>
+        </div>
+      </div>
+    </MiniPanel>
+  </div>
+  <div class="timetable">
+    <table>
+      <tbody>
+        <tr>
+          <td class="tt-time">16:00</td>
+          <td class="tt-event">クリエイティブコーディングワークショップ</td>
+        </tr>
+        <tr>
+          <td class="tt-time">17:45</td>
+          <td class="tt-event">Door Open</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>18:00</p>
+            <p>-</p>
+            <p>18:10</p>
+          </td>
+          <td class="tt-event">Opening</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>18:10</p>
+            <p>-</p>
+            <p>18:25</p>
+          </td>
+          <td class="tt-event">クリエイティブコーディング発表会</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>18:25</p>
+            <p>-</p>
+            <p>18:30</p>
+          </td>
+          <td class="tt-talk">1区 5m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>18:30</p>
+            <p>-</p>
+            <p>18:50</p>
+          </td>
+          <td class="tt-talk">2区 20m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>18:50</p>
+            <p>-</p>
+            <p>19:00</p>
+          </td>
+          <td class="tt-talk">3区 10m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>19:00</p>
+            <p>-</p>
+            <p>19:05</p>
+          </td>
+          <td class="tt-talk">4区 5m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>19:05</p>
+            <p>-</p>
+            <p>19:25</p>
+          </td>
+          <td class="tt-talk">5区 20m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>19:25</p>
+            <p>-</p>
+            <p>19:40</p>
+          </td>
+          <td class="tt-event">Break</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>19:40</p>
+            <p>-</p>
+            <p>20:00</p>
+          </td>
+          <td class="tt-talk">6区 20m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>20:00</p>
+            <p>-</p>
+            <p>20:05</p>
+          </td>
+          <td class="tt-talk">7区 5m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>20:05</p>
+            <p>-</p>
+            <p>20:15</p>
+          </td>
+          <td class="tt-talk">8区 15m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>20:15</p>
+            <p>-</p>
+            <p>20:35</p>
+          </td>
+          <td class="tt-talk">9区 20m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>20:35</p>
+            <p>-</p>
+            <p>20:40</p>
+          </td>
+          <td class="tt-talk">10区 5m</td>
+        </tr>
+        <tr>
+          <td class="tt-timerange">
+            <p>20:40</p>
+            <p>-</p>
+            <p>20:50</p>
+          </td>
+          <td class="tt-event">Closing</td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
+</div>
